+++ app/institutions/dashboard/-components/departments-panel/template.hbs
@@ -1,0 +1,1 @@
+{{ember-chart type='pie' data=this.chartData options=this.chartOptions height=150}}
