+++ app/institutions/dashboard/-components/projects-panel/template.hbs
@@ -1,0 +1,8 @@
+{{ember-chart type='pie' data=this.chartData options=this.chartOptions height=150}}
+<h3>{{this.numProjects}}</h3>
+<p>
+    <strong>{{this.institution.statSummary.numPublicProjects}}</strong> {{t 'institutions.dashboard.public'}}
+</p>
+<p>
+    <strong>{{this.institution.statSummary.numPrivateProjects}}</strong> {{t 'institutions.dashboard.private'}}
+</p>
